requirements.txt
pyTelegramBotApi
psycopg2
flask
requests
yandex-money-sdk
blockchain
coinbase